+++ code/core/trkCore.h
@@ -98,7 +98,7 @@ std::vector<std::vector<float>>&    out_eta_vec,
 std::vector<std::vector<float>>&    out_etaErr_vec,
 std::vector<std::vector<float>>&    out_phi_vec,
 std::vector<std::vector<int>>&    out_superbin_vec,
-std::vector<std::vector<int>>&    out_pixelType_vec,
+std::vector<std::vector<int8_t>>&    out_pixelType_vec,
 std::vector<std::vector<short>>&    out_isQuad_vec);
 //float addInputsToEventPreLoad(SDL::Event& event, bool useOMP,std::vector<float> trkX,std::vector<float> trkY,std::vector<float> trkZ,
 float addInputsToEventPreLoad(SDL::Event* event, bool useOMP,std::vector<float> trkX,std::vector<float> trkY,std::vector<float> trkZ,
@@ -118,7 +118,7 @@ std::vector<float>    eta_vec,
 std::vector<float>    etaErr_vec,
 std::vector<float>    phi_vec,
 std::vector<int>    superbin_vec,
-std::vector<int>    pixelType_vec,
+std::vector<int8_t>    pixelType_vec,
 std::vector<short>   isQuad_vec);
 
 TVector3 calculateR3FromPCA(const TVector3& p3, const float dxy, const float dz);
